# [CI] Speed up check_images_exist