@supertopher 2022-11-02
ran locally.  output for others
```shell
√ airbyte $ ./tools/bin/check_images_exist.sh all
./tools/bin/check_images_exist.sh:20: set +o xtrace
checking images for: all
Checking platform images exist...
	airbyte/init: 0.40.18
	URL: https://hub.docker.com/v2/repositories/airbyte/init/tags/0.40.18
	STATUS: found

	airbyte/bootloader: 0.40.18
	URL: https://hub.docker.com/v2/repositories/airbyte/bootloader/tags/0.40.18
	STATUS: found

	airbyte/db: 0.40.18
	URL: https://hub.docker.com/v2/repositories/airbyte/db/tags/0.40.18
	STATUS: found

	airbyte/worker: 0.40.18
	URL: https://hub.docker.com/v2/repositories/airbyte/worker/tags/0.40.18
	STATUS: found

	airbyte/server: 0.40.18
	URL: https://hub.docker.com/v2/repositories/airbyte/server/tags/0.40.18

<more lines here>


	airbyte/source-firebolt: 0.1.0
	URL: https://hub.docker.com/v2/repositories/airbyte/source-firebolt/tags/0.1.0
	STATUS: found

	airbyte/source-elasticsearch: 0.1.0
	URL: https://hub.docker.com/v2/repositories/airbyte/source-elasticsearch/tags/0.1.0
	STATUS: found

	airbyte/source-waiteraid: 0.1.0
	URL: https://hub.docker.com/v2/repositories/airbyte/source-waiteraid/tags/0.1.0
	STATUS: found

	airbyte/source-yandex-metrica: 0.1.0
	URL: https://hub.docker.com/v2/repositories/airbyte/source-yandex-metrica/tags/0.1.0
	STATUS: found

	airbyte/source-zoom: 0.1.0
	URL: https://hub.docker.com/v2/repositories/airbyte/source-zoom/tags/0.1.0
	STATUS: found

Success! All connector images exist!
Image check complete.
```